requirements.txt
pymongo>=4.6.0
python-dotenv>=1.0.0
openai>=1.12.0
langchain>=0.1.0
langchain-openai>=0.0.8
langchain-core>=0.1.0
boto3>=1.34.0
botocore>=1.34.0

huggingface-hub>=0.20.0
requests>=2.31.0

numpy>=1.24.0
pandas>=2.0.0

pathlib2>=2.3.0

fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
